Extract PatchOverview and ShaderInstanceEditor from PatchHolderEditor.
