+++ src/jsMain/kotlin/baaahs/ui/PatchHolderEditor.kt
@@ -3,7 +3,6 @@ package baaahs.ui
 import baaahs.app.ui.PatchHolderEditorHelpText
 import baaahs.app.ui.appContext
 import baaahs.show.ShaderChannel
-import baaahs.show.ShaderType
 import baaahs.show.mutable.MutablePatch
 import baaahs.show.mutable.MutablePatchHolder
 import baaahs.show.mutable.MutableShaderInstance
@@ -14,39 +13,26 @@ import kotlinx.html.js.onChangeFunction
 import kotlinx.html.js.onClickFunction
 import kotlinx.html.js.onSubmitFunction
 import materialui.*
-import materialui.components.avatar.avatar
 import materialui.components.breadcrumbs.breadcrumbs
 import materialui.components.button.button
 import materialui.components.button.enums.ButtonColor
-import materialui.components.card.card
-import materialui.components.cardheader.cardHeader
 import materialui.components.container.container
 import materialui.components.dialogactions.dialogActions
 import materialui.components.dialogcontent.dialogContent
 import materialui.components.dialogtitle.dialogTitle
-import materialui.components.divider.divider
 import materialui.components.drawer.drawer
 import materialui.components.drawer.enums.DrawerAnchor
 import materialui.components.drawer.enums.DrawerStyle
 import materialui.components.drawer.enums.DrawerVariant
 import materialui.components.formcontrol.enums.FormControlVariant
-import materialui.components.formcontrol.formControl
-import materialui.components.formhelpertext.formHelperText
 import materialui.components.link.link
 import materialui.components.list.enums.ListStyle
 import materialui.components.list.list
 import materialui.components.listitem.listItem
 import materialui.components.listitemicon.listItemIcon
 import materialui.components.listitemtext.listItemText
-import materialui.components.menuitem.menuItem
-import materialui.components.paper.enums.PaperStyle
 import materialui.components.portal.portal
-import materialui.components.select.select
 import materialui.components.textfield.textField
-import materialui.components.typography.enums.TypographyColor
-import materialui.components.typography.enums.TypographyDisplay
-import materialui.components.typography.enums.TypographyVariant
-import materialui.components.typography.typography
 import materialui.components.typography.typographyH1
 import materialui.components.typography.typographyH6
 import org.w3c.dom.HTMLInputElement
@@ -56,7 +42,6 @@ import react.dom.div
 import react.dom.form
 import react.dom.header
 import styled.StyleSheet
-import styled.styledDiv
 
 @Suppress("UNCHECKED_CAST")
 fun <T> Event.targetEl(): T = target as T
@@ -77,9 +62,14 @@ val PatchHolderEditor = xComponent<PatchHolderEditorProps>("PatchHolderEditor") 
     }
     val handleChange = useCallback { forceRender() }
 
+    var visiblePanel by state { 0 }
+
     var selectedPatch by state { props.mutablePatchHolder.patches.firstOrNull() }
     var selectedShaderInstance by state<MutableShaderInstance?> { null }
-    var visiblePanel by state { 0 }
+    val handleSelectShader = useCallback { selected: MutableShaderInstance ->
+        selectedShaderInstance = selected
+        visiblePanel = 1
+    }
 
     val x = this
     val styles = PatchHolderEditorStyles
@@ -130,80 +120,12 @@ val PatchHolderEditor = xComponent<PatchHolderEditorProps>("PatchHolderEditor") 
                 }
 
                 div(+styles.patchOverview) {
-                    val mutablePatch = selectedPatch
-                    if (mutablePatch == null) {
+                    if (selectedPatch == null) {
                         typographyH6 { +"No patch selected." }
                     } else {
-                        mutablePatch.mutableShaderInstances.forEachIndexed { index, mutableShaderInstance ->
-                            val shader = mutableShaderInstance.mutableShader
-
-                            card(+styles.shaderCard on PaperStyle.root) {
-                                attrs.onClickFunction = x.eventHandler("handleShaderInstanceClick-$index") {
-                                    selectedShaderInstance = selectedPatch!!.mutableShaderInstances[index]
-                                    visiblePanel = 1
-                                }
-
-                                cardHeader {
-                                    attrs.avatar {
-                                        avatar { icon(Icons.forShader(shader.type)) }
-                                    }
-                                    attrs.title { +shader.title }
-//                                attrs.subheader { +"${shader.type.name} Shader" }
-                                }
-
-                                shaderPreview {
-                                    attrs.mutableShaderInstance = mutableShaderInstance
-                                    attrs.width = styles.cardWidth
-                                    attrs.height = styles.cardWidth
-                                }
-
-                                div(+styles.shaderCardContent) {
-                                    typography {
-                                        attrs.display = TypographyDisplay.block
-                                        attrs.variant = TypographyVariant.body2
-                                        attrs.color = TypographyColor.textSecondary
-                                        +"${shader.type.name} Shader"
-                                    }
-                                }
-//                            cardActions {
-//                                button { +"Edit" }
-//                            }
-                            }
-                        }
-
-                        card {
-                            menuButton {
-                                attrs.icon = AddCircleOutline
-                                attrs.label = "New Shader…"
-
-                                attrs.items = ShaderType.values().map { type ->
-                                    MenuItem("New ${type.name} Shader…") {
-                                        val newShader = type.shaderFromTemplate().build()
-                                        val contextShaders =
-                                            mutablePatch.mutableShaderInstances.map { it.mutableShader.build() } + newShader
-                                        val unresolvedPatch = appContext.autoWirer.autoWire(
-                                            *contextShaders.toTypedArray(),
-                                            focus = newShader
-                                        )
-                                        mutablePatch.addShaderInstance(newShader) {
-                                            // TODO: Something better than this.
-                                            val resolved = unresolvedPatch
-                                                .acceptSymbolicChannelLinks()
-                                                .resolve()
-                                                .mutableShaderInstances[0]
-                                            incomingLinks.putAll(resolved.incomingLinks)
-                                            shaderChannel = resolved.shaderChannel
-                                        }
-                                        this@xComponent.forceRender()
-                                    }
-                                } + MenuItem("Import…") { }
-                            }
-
-                            typography {
-                                attrs.display = TypographyDisplay.block
-                                attrs.variant = TypographyVariant.subtitle1
-                                +"New Shader…"
-                            }
+                        patchOverview {
+                            attrs.mutablePatch = selectedPatch!!
+                            attrs.onSelectShaderInstance = handleSelectShader
                         }
                     }
                 }
@@ -214,96 +136,14 @@ val PatchHolderEditor = xComponent<PatchHolderEditorProps>("PatchHolderEditor") 
     val panels = arrayListOf(patchNavPanel)
 
     val shaderPanel = selectedShaderInstance?.let { shaderInstance ->
-        val shader = shaderInstance.mutableShader.build()
+        val shader = shaderInstance.mutableShader
         val shaderChannels = props.mutablePatchHolder.findShaderChannels() + ShaderChannel.Main
-        val handleUpdate =
-            handler("handleShaderUpdate", shaderInstance) { block: MutableShaderInstance.() -> Unit ->
-                selectedShaderInstance!!.block()
-                forceRender()
-            }
-
         Panel(shader.title, Icons.forShader(shader.type)) {
-            styledDiv {
-                css.display = Display.grid
-                css.gridTemplateColumns = GridTemplateColumns("auto min-content auto")
-                css.gap = Gap(1.em.toString())
-
-                div {
-                    linksEditor {
-                        attrs.mutablePatch = selectedPatch!!
-                        attrs.showBuilder = showBuilder
-                        attrs.mutableShaderInstance = shaderInstance
-                        attrs.shaderChannels = shaderChannels
-                        attrs.onChange = handleChange
-                    }
-                }
-
-                div {
-                    shaderPreview {
-                        attrs.mutableShaderInstance = shaderInstance
-                        attrs.width = 250.px
-                        attrs.height = 250.px
-                    }
-                }
-
-                div {
-                    typographyH6 { +"Meta and gadgets and stuff!" }
-
-                    div(+Styles.shaderMeta) {
-                        formControl {
-                            textField {
-                                attrs.autoFocus = false
-                                attrs.fullWidth = true
-                                attrs.value = shaderInstance.mutableShader.title
-                                attrs.onChangeFunction = { event: Event ->
-                                    val str = event.target!!.asDynamic().value as String
-                                    handleUpdate { mutableShader.title = str }
-                                }
-                            }
-                            formHelperText { +"Shader Name" }
-                        }
-
-                        formControl {
-                            select {
-                                attrs.value(shaderInstance.shaderChannel?.id ?: "")
-                                attrs.onChangeFunction = { event: Event ->
-                                    val channelId = event.target!!.asDynamic().value as String
-                                    handleUpdate { shaderChannel = if (channelId.isNotBlank()) ShaderChannel(channelId) else null }
-                                }
-                                shaderChannels.sortedBy { it.id }.forEach { shaderChannel ->
-                                    menuItem {
-                                        attrs["value"] = shaderChannel.id
-                                        +shaderChannel.id
-                                    }
-                                }
-
-                                divider {}
-                                menuItem {
-                                    attrs["value"] = ""
-                                    +"None"
-                                }
-                            }
-                            formHelperText { +"Channel" }
-                        }
-
-                        formControl {
-                            textField {
-                                attrs.value = shaderInstance.priority
-                                attrs.onChangeFunction = { event: Event ->
-                                    val priorityStr = event.target!!.asDynamic().value as String
-                                    handleUpdate { priority = priorityStr.toFloat() }
-                                }
-                            }
-                            formHelperText { +"Priority" }
-                        }
-                    }
-
-                }
-            }
-
-            shaderEditor {
+            shaderInstanceEditor {
+                attrs.mutablePatch = selectedPatch!!
                 attrs.mutableShaderInstance = shaderInstance
                 attrs.shaderChannels = shaderChannels
+                attrs.showBuilder = showBuilder
                 attrs.onChange = handleChange
             }
         }
